moved ServiceResponse to types dir
