+++ src/commands/create-issue.ts
@@ -4,7 +4,7 @@ import fs from "node:fs";
 import CreateIssueOpt from "./opts/create-issue-opt.js";
 import {SleekCommand} from "../sleek-command.js";
 import {createIssue, validateInputFileSchema} from "../service/create-issue-svc.js";
-import {issueData} from "../types/issue.js";
+import {IssueData} from "../types/issue.js";
 
 const Ajv = _Ajv as unknown as typeof _Ajv.default;
 
@@ -31,7 +31,7 @@ export default class CreateIssue extends SleekCommand {
         // validateChart(helmCharName, helmChartVersion);
 
         // create issue base in the file input
-        const title = `Onboarding ${(data.body as issueData).sellerMarketPlaceAlias} ${(data.body as issueData).addon.name}@${(data.body as issueData).addon.version}`;
+        const title = `Onboarding ${(data.body as IssueData).sellerMarketPlaceAlias} ${(data.body as IssueData).addon.name}@${(data.body as IssueData).addon.version}`;
         const body= `Issue body:\n\n\`\`\`yaml\n${fileContents}\`\`\`\n`;
         const createIssueResponse = await createIssue(title, body,this, ['pending'])
