+++ src/services/base-service.ts
@@ -2,6 +2,7 @@
  * Is the base class of all our services, holds the configs
  * that are global to services
  */
+import {ServiceResponse} from "../types/service.js";
 
 export interface ServiceConfig {
   // defines chart location
@@ -10,17 +11,6 @@ export interface ServiceConfig {
   // define other stuff common to all services here
 }
 
-export type ServiceResponse<T> = {
-  success: boolean,
-  body?: T,
-  error?: {
-    input: Error | string,
-    options?: {
-      code?: string;
-      exit?: number;
-    }
-  }
-}
 
 export abstract class BaseService {
   protected config: ServiceConfig;
