+++ src/services/helm.ts
@@ -1,4 +1,5 @@
-import {BaseService, ServiceResponse} from "./base-service.js";
+import {BaseService} from "./base-service.js";
+import {ServiceResponse} from "../types/service.js";
 
 export default class HelmManagerService extends BaseService {
   public static async pullAndUnzipChart() {
